GPG copy and paste issues
There were references of SIFLaanseTacticalMissile01 XSL0111 left in the original file. I had only adjusted the turn rates from 10 and 12 to be at least 20. I changed the incorrect commented out references to SIFLaanseTacticalMissileSCU and XSL0301, but left the other 2 occurrences in the code alone based on the saying, if it ain't broke, don't fix it.